sol2: Use appropriate values objects for the various C and C++ standards
I'm reasonable confident of the C cases, less confident of the C++
cases.

This is on the basis that c99 and greater implies XPG6 (and vice versa), as the
manual pages and headers say/enforce.  It is possible that C11 should
imply xpg7 if and when we support that.

C99 and greater get strict ansi.

C otherwise gets transitional ansi (and no UNIX standard)

C++ gets strict ansi, other than in C++98.  And gets no UNIX standards
objects at all.